jets: reorganize jets
